+++ packages/plugin-packages/mashroom-portal-remote-app-registry/src/js/webapp/mashroom-bootstrap-webapp.js
@@ -1,13 +1,16 @@
 // @flow
 
+import context from '../context';
 import webapp from './webapp';
 
 import type {
     MashroomWebAppPluginBootstrapFunction
 } from '@mashroom/mashroom/type-definitions';
 
+const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig) => {
+    const { showAddRemoteAppForm } = pluginConfig;
+    context.webUIShowAddRemoteAppForm = showAddRemoteAppForm;
 
-const bootstrap: MashroomWebAppPluginBootstrapFunction = async () => {
     return webapp;
 };