+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Remote App Registry: Added plugin config property to hide the *Add a new Remote Portal App Endpoint* form from the Admin UI
  * Remote App Registry: Moved config properties from the *Mashroom Portal Remote App Registry Webapp* plugin to the
    *Mashroom Portal Remote App Registry* plugin where it belongs (**BREAKING CHANGE**)